FEAT: smart contract model player beast

+++ dojo/src/models.cairo
@@ -4,7 +4,7 @@ use starknet::ContractAddress;
 #[dojo::model]
 pub struct Beast {
     #[key]
-    pub beast_id: u32,
+    pub beast_id: u128,
     pub player: ContractAddress,
     pub specie: u32,
     pub is_alive: bool,
